% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloch.r
\name{bloch.image}
\alias{bloch.image}
\title{Forward Bloch Equation (for whole MR image):}
\usage{
bloch.image(W, TE, TR)
}
\arguments{
\item{W}{A numeric matrix of size n x 3}

\item{TE}{A numeric vector, TE values for the training set}

\item{TR}{A numeric vector, TR values for the training set}
}
\value{
The matrix corresponding to the MR signal values Bloch equation predicts(i.e., \eqn{\hat{\nu}})
}
\description{
Forward Bloch Equation (for whole MR image):
}
\examples{
## Sample row of parameters, W
W <- rbind(c(50, 0.01, 0.003), c(36, 0.02, 0.04)) ## Two sample rows
## Design parameters
TE <- c(0.01, 0.03, 0.04, 0.01)
TR <- c(0.6, 0.6, 1, 0.8)
## Forward transformed values:
bloch.image(W, TE, TR)
}
