% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloch.r
\name{bloch}
\alias{bloch}
\title{Forward Bloch Equation (for a single voxel):}
\usage{
bloch(w, TE, TR)
}
\arguments{
\item{w}{A numeric Vector of size 3}

\item{TE}{A numeric vector, TE values for the training set}

\item{TR}{A numeric vector, TR values for the training set}
}
\value{
The vector corresponding to the MR signal values Bloch equation predicts(i.e., \eqn{\hat{\nu}})
}
\description{
Forward Bloch Equation (for a single voxel):
}
\examples{
## Sample row of parameters, w
w <- c(50, 0.01, 0.003)
## Design parameters
TE <- c(0.01, 0.03, 0.04, 0.01)
TR <- c(0.6, 0.6, 1, 0.8)
## Forward transformed values: 
bloch(w, TE, TR)
}
