% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Bloch_eqn}
\alias{Bloch_eqn}
\title{Forward Bloch Equation (for a single voxel):}
\usage{
Bloch_eqn(W_row, TE, TR)
}
\description{
Forward Bloch Equation (for a single voxel):
}
\examples{
## Sample row of parameters, W
W_row <- c(50, 0.01, 0.003)
## Design parameters
TE <- c(0.01, 0.03, 0.04, 0.01)
TR <- c(0.6, 0.6, 1, 0.8)
## Forward transformed values: 
Bloch_eqn(W_row, TE, TR)
}
