% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Bloch_eqn_R}
\alias{Bloch_eqn_R}
\title{Forward Bloch Equation (for a single voxel):}
\usage{
Bloch_eqn_R(W_row, TE, TR)
}
\arguments{
\item{W_row}{A numeric Vector of size 3, any row of W}

\item{TE}{A numeric vector, TE values for the training set}

\item{TR}{A numeric vector, TR values for the training set}
}
\value{
The vector corresponding to the MR signal values Bloch equation predicts(i.e., \eqn{\hat{\nu}})
}
\description{
Forward Bloch Equation (for a single voxel):
}
\examples{
## Sample row of parameters, W
W_row <- c(50, 0.01, 0.003)
## Design parameters
TE <- c(0.01, 0.03, 0.04, 0.01)
TR <- c(0.6, 0.6, 1, 0.8)
## Forward transformed values: 
Bloch_eqn_R(W_row, TE, TR)
}
