% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dee_2_v_ij_dee_W_ik_dee_W_ik1}
\alias{dee_2_v_ij_dee_W_ik_dee_W_ik1}
\title{2nd Derivative corresponding to forward Bloch Equation (for a single voxel):}
\usage{
dee_2_v_ij_dee_W_ik_dee_W_ik1(W_row, TE, TR, j, k, k1)
}
\arguments{
\item{W_row}{A numeric Vector of size 3, any row of W}

\item{TE}{A numeric vector, TE values for the training set}

\item{TR}{A numeric vector, TR values for the training set}

\item{j}{index of image number or corresponding settings number}

\item{k, }{k1  co-ordinate of the W(i.e., \eqn{W_1} or \eqn{W_2} or \eqn{W_3} ) with respect to which the derivative is taken}
}
\value{
The vector corresponding to derivatives Bloch equation predicts(i.e., \eqn{\hat{\nu}})
}
\description{
2nd Derivative corresponding to forward Bloch Equation (for a single voxel):
}
\examples{
## Sample row of parameters, W
W_row <- c(50, 0.01, 0.003)
## Design parameters
TE <- c(0.01, 0.03, 0.04, 0.01)
TR <- c(0.6, 0.6, 1, 0.8)
## Forward transformed values: 
dee_2_v_ij_dee_W_ik_dee_W_ik1(W_row, TE, TR, 1, 1, 2)	## d^2v_i1/(dW_i1 dW_i2)
}
