% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dee_v_ij_dee_W_ik}
\alias{dee_v_ij_dee_W_ik}
\title{Derivative corresponding to forward Bloch Equation (for a single voxel):}
\usage{
dee_v_ij_dee_W_ik(W_row, TE, TR, j, k)
}
\arguments{
\item{W_row}{A numeric Vector of size 3, any row of W}

\item{TE}{A numeric vector, TE values for the training set}

\item{TR}{A numeric vector, TR values for the training set}

\item{j}{index of image number or corresponding settings number}

\item{k}{co-ordinate of the W(i.e., \eqn{W_1} or \eqn{W_2} or \eqn{W_3} ) with respect to which the derivative is taken}
}
\value{
The vector corresponding to derivatives Bloch equation predicts(i.e., \eqn{\hat{\nu}})
}
\description{
Derivative corresponding to forward Bloch Equation (for a single voxel):
}
\examples{
## Sample row of parameters, W
W_row <- c(50, 0.01, 0.003)
## Design parameters
TE <- c(0.01, 0.03, 0.04, 0.01)
TR <- c(0.6, 0.6, 1, 0.8)
## Forward transformed values: 
dee_v_ij_dee_W_ik(W_row, TE, TR, 1, 1)					## dv_i1/dW_i1
}
