% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma.R
\name{estimate.sigma.j}
\alias{estimate.sigma.j}
\title{Estimating rice noise parameter from a MR image using mixture of rice distribution.}
\usage{
estimate.sigma.j(
  r.col,
  min_grp = 5L,
  max_grp = 15L,
  init_iter = 4L,
  EM_iter = 15L,
  eps = 1e-04
)
}
\arguments{
\item{r.col}{An MR image in a vector format}

\item{min_grp}{Minimum number of possible mixture groups for the Rice-mixture distribution}

\item{max_grp}{Maximum number of possible mixture groups for the Rice-mixture distribution}

\item{init_iter}{Maximum iteration number for initial value estimation}

\item{EM_iter}{Maximum iteration number for EM estimation}

\item{eps}{Relative error stopping criteria for the EM algorithm}
}
\value{
The estimate of rice noise parameter for the image \code{r.col}
}
\description{
Estimating rice noise parameter from a MR image using mixture of rice distribution.
}
\examples{
\donttest{
## Sample row of image,
file_name <- system.file("extdata", "new_phantom.nii.gz", package = "symR", mustWork = TRUE)
phantom <- RNifti::readNifti(file_name, internal = TRUE)
phantom <- apply(phantom, 4, function(x) {
  c(x)
})
phantom[phantom == 0.0] <- 0.5 ## Pre-processing to remove the -Inf issue in likelihood.
estimate.sigma.j(phantom[, 1])
}
}
