% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloch.r
\name{Generate.noisy.image}
\alias{Generate.noisy.image}
\title{Forward Bloch Equation (for whole MR image):}
\usage{
Generate.noisy.image(W, TE, TR, sigma)
}
\arguments{
\item{W}{A numeric matrix of size n x 3}

\item{TE}{A numeric vector, TE values}

\item{TR}{A numeric vector, TR values}

\item{sigma}{A numeric vector, Rice noise parameter values}
}
\value{
The matrix corresponding to the Rice noise contaminated MR signal values Bloch equation predicts
}
\description{
Forward Bloch Equation (for whole MR image):
}
\examples{
## Sample row of parameters, W
W <- rbind(c(50, 0.01, 0.003), c(36, 0.02, 0.04)) ## Two sample rows
## Design parameters
TE <- c(0.01, 0.03, 0.04, 0.01)
TR <- c(0.6, 0.6, 1, 0.8)
sig <- c(1.2, 2.1, 1.8)
## Forward transformed values:
Generate.noisy.image(W, TE, TR, sigma)
}
