% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.var.R
\name{prediction.var}
\alias{prediction.var}
\title{Variance estimates corresponding to contrasts using Information Matrix and Delta method}
\usage{
prediction.var(
  W,
  Psi.inv,
  betavec,
  contrast,
  dimen,
  TE.train,
  TR.train,
  sigma.train,
  train,
  TE.test,
  TR.test,
  sigma.test,
  test,
  train.scale,
  TE.scale,
  TR.scale,
  black.list,
  cg_maxiter = 50L,
  cg_tol = 1e-06,
  penalized = 1L
)
}
\arguments{
\item{W}{A numeric Matrix of size n x 3 - the estimate of primary parameters.}

\item{Psi.inv}{A numeric Matrix of size 3 x 3 - estimate of Psi inverse, corresponding to the correlation matrix between columns of the primary parameters W}

\item{betavec}{A numeric vector of size 3 - estimate of the strength of spatial factor in each directions.}

\item{contrast}{A sparse vector corresponding to the contrast of which variance is to be estimated.}

\item{dimen}{The dimension of the train MR signals (possibly read from the nifti file, the format would be: )}

\item{TE.train}{A numeric vector, TE values for the training set}

\item{TR.train}{A numeric vector, TR values for the training set}

\item{sigma.train}{A numeric vector, sigma_j values for the training set}

\item{train}{A numeric matrix of size n x m of observed signals to be trained on, where m is the number of training MR images.}

\item{TE.test}{A numeric vector, TE values for the test set}

\item{TR.test}{A numeric vector, TR values for the test set}

\item{sigma.test}{A numeric vector, sigma_j values for the test set}

\item{test}{A numeric Matrix, supplied as the values of the voxels of size n x m}

\item{train.scale}{A positive real number by which voxels are scaled}

\item{TE.scale}{A positive real number by which TE values are scaled}

\item{TR.scale}{A positive real number by which TR values are scaled}

\item{black.list}{A numeric vector of size n signifying the background voxels}

\item{cg_maxiter}{Maximum iteration number of conjugate-gradient method}

\item{cg_tol}{tolerance number of conjugate-gradient method}

\item{penalized}{This is 1 if penalized methods are used(such as "AECM", "OSL").}
}
\value{
The vector corresponding to variance estimate
}
\description{
Variance estimates corresponding to contrasts using Information Matrix and Delta method
}
\examples{
W <- rbind(c(40, 0.01, 0.003), c(36, 0.02, 0.04)) ## Two sample rows of parameters, W
test <- rbind(c(56, 52, 57, 51), c(39, 37, 33, 34.4))

## Test design parameters
TE <- c(0.01, 0.03, 0.04, 0.01)
TR <- c(0.6, 0.6, 1, 0.8)
sig <- c(1.2, 1.1, 1.4, 1.2)
mask <- c(0, 0)
}
