% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{v_mat_R}
\alias{v_mat_R}
\title{Forward Bloch Equation for whole MR image:}
\usage{
v_mat_R(W, TE, TR)
}
\arguments{
\item{W}{A numeric Matrix of size n x 3 corresponding to the transformed parameters, \eqn{\rho, \mathrm{T}_1, \mathrm{T}_2}}

\item{TE}{A numeric vector, TE values for the training set}

\item{TR}{A numeric vector, TR values for the training set}
}
\value{
The matrix corresponding to the MR signal values Bloch equation for a whole image
}
\description{
Forward Bloch Equation for whole MR image:
}
\examples{
## Sample row of parameters, W
W <- rbind(c(50, 0.01, 0.003), c(36, 0.02, 0.04))	## Two sample rows
## Design parameters
TE <- c(0.01, 0.03, 0.04, 0.01)
TR <- c(0.6, 0.6, 1, 0.8)
## Forward transformed values: 
v_mat_R(W, TE, TR)
}
